initial move of filer_metadata to generic metadata
